Updating default currency rates using ExchangeRate API implemented

+++ eShop.Distribution/Entities/DefaultCurrencyRate.cs
@@ -4,7 +4,7 @@ namespace eShop.Distribution.Entities
 {
     public class DefaultCurrencyRate : EntityBase, ICurrencyRate
     {
-        public double Rate { get; set; }
+        public float Rate { get; set; }
         public EmbeddedCurrency SourceCurrency { get; set; }
         public EmbeddedCurrency TargetCurrency { get; set; }
         public DateTimeOffset CreatedAt { get; set; } = DateTimeOffset.UtcNow;
