+++ eShop.Distribution/Entities/ICurrencyRate.cs
@@ -2,7 +2,7 @@
 {
     public interface ICurrencyRate
     {
-        public double Rate { get; set; }
+        public float Rate { get; set; }
         public EmbeddedCurrency SourceCurrency { get; set; }
         public EmbeddedCurrency TargetCurrency { get; set; }
     }
