+++ eShop.Distribution/Entities/UserCurrencyRate.cs
@@ -3,7 +3,7 @@
     public class UserCurrencyRate : ICurrencyRate
     {
         public Guid Id { get; set; } = Guid.NewGuid();
-        public double Rate { get; set; }
+        public float Rate { get; set; }
         public EmbeddedCurrency SourceCurrency { get; set; }
         public EmbeddedCurrency TargetCurrency { get; set; }
     }
